backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Redis
redis[hiredis]==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Background tasks
arq==0.26.1

# PDF generation
weasyprint==63.1
jinja2==3.1.5

# HTTP client (for GitHub, Discord)
httpx==0.28.1

# HTML sanitization
nh3==0.2.20

# Rate limiting
python-multipart==0.0.20

# Logging
structlog==24.4.0

# Utils
python-slugify==8.0.4
